+++ src/map/chrif.cpp
@@ -1349,7 +1349,7 @@ int chrif_save_scdata(struct map_session_data *sd) { //parses the sc_data of the
 	}
 
 	WFIFOW(char_fd,12) = count;
-	WFIFOW(char_fd,2) = 14 +count*sizeof(struct status_change_data); //Total packet size
+	WFIFOW(char_fd,2) = (int)(14 + count * (uint16)sizeof(struct status_change_data)); //Total packet size
 	WFIFOSET(char_fd,WFIFOW(char_fd,2));
 #endif
 	return 0;
@@ -1390,7 +1390,7 @@ int chrif_skillcooldown_save(struct map_session_data *sd) {
 		return 0;
 
 	WFIFOW(char_fd, 12) = count;
-	WFIFOW(char_fd, 2) = 14 + count * sizeof (struct skill_cooldown_data);
+	WFIFOW(char_fd, 2) = 14 + count * (uint16)sizeof(struct skill_cooldown_data);
 	WFIFOSET(char_fd, WFIFOW(char_fd, 2));
 
 	return 0;
