+++ src/map/mob.cpp
@@ -2854,7 +2854,7 @@ int mob_dead(struct mob_data *md, struct block_list *src, int type)
 				char message[128];
 				sprintf (message, msg_txt(NULL,541), mvp_sd->status.name, md->name, it->ename.c_str(), (float)drop_rate/100);
 				//MSG: "'%s' won %s's %s (chance: %0.02f%%)"
-				intif_broadcast(message,strlen(message)+1,BC_DEFAULT);
+				intif_broadcast(message,(int)(strlen(message)+1),BC_DEFAULT);
 			}
 			// Announce first, or else ditem will be freed. [Lance]
 			// By popular demand, use base drop rate for autoloot code. [Skotlex]
@@ -3024,7 +3024,7 @@ int mob_dead(struct mob_data *md, struct block_list *src, int type)
 					char message[128];
 					sprintf (message, msg_txt(NULL,541), mvp_sd->status.name, md->name, i_data->ename.c_str(), temp/100.);
 					//MSG: "'%s' won %s's %s (chance: %0.02f%%)"
-					intif_broadcast(message,strlen(message)+1,BC_DEFAULT);
+					intif_broadcast(message,(int)(strlen(message)+1),BC_DEFAULT);
 				}
 
 				mob_setdropitem_option(&item, &mdrop[i]);
@@ -6402,8 +6402,8 @@ static void mob_load(void)
 	mob_chat_db.load();	// load before mob_skill_db
 
 	for(int i = 0; i < ARRAYLENGTH(dbsubpath); i++){	
-		int n1 = strlen(db_path)+strlen(dbsubpath[i])+1;
-		int n2 = strlen(db_path)+strlen(DBPATH)+strlen(dbsubpath[i])+1;
+		int n1 = (int)(strlen(db_path)+strlen(dbsubpath[i])+1);
+		int n2 = (int)(strlen(db_path)+strlen(DBPATH)+strlen(dbsubpath[i])+1);
 
 		char* dbsubpath1 = (char*)aMalloc(n1+1);
 		char* dbsubpath2 = (char*)aMalloc(n2+1);
